+++ memorybound/memory.cpp
@@ -1,6 +1,6 @@
 #include "memory.h"
 #include "memory_kernel.h"
 
-void multiMatrix(float *Ma, float *Mb, float *Mc) {
+void multiMatrix(double *Ma, double *Mb, double *Mc) {
   matrixMulti_caller(Ma, Mb, Mc);
 }
